Finished

+++ exercises/loopArray/loopArray.cpp
@@ -4,7 +4,7 @@
 int main()
 {
     constexpr int array[]{ 4, 6, 7, 3, 8, 2, 1, 9, 5 };
-    int arraySize{ std::size(array) };
+    int arraySize{ static_cast<int>(std::size(array)) };
 
     for (int i{ 0 }; i < arraySize; ++i)
         std::cout << array[i] << '\n';
